chore(adapter): rename integrator submodule to adapter
